requirements.txt
langchain==0.1.16
faiss-cpu==1.8.0
sentence-transformers==2.7.0
streamlit==1.35.0
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
langchain-google-genai==0.1.3
pandas==2.2.2